+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## [Unreleased]
 
+### Added
+- `check-http`: Added ability to use username and password in proxy url (@mclarkson)
+
 ## [2.6.0] - 2017-07-31
 ### Added
 - ruby 2.4 testing (@majormoses)